robbin

+++ news/news.php
@@ -146,7 +146,7 @@
 	   		if( ($rs_num-5) % $page_size ){                                  
 	       		$page_count = (int)($rs_num / $page_size) + 2;           
 	   		}else{
-	       		$page_count = $rs_num / $page_size + 1;                      
+	       		$page_count = (int)($rs_num / $page_size) + 1;                      
 	  		}
 		}
 		else{
